Mutualize code between Scatter and Curve in a Points base class
